Update the hello world example.
